# openapi/v1/baselines/1.0.0.openapi.json
{
  "components": {
    "schemas": {
      "ActTaskDto": {
        "properties": {
          "action": {
            "enum": [
              "accept",
              "complete",
              "approve",
              "reopen",
              "close"
            ],
            "example": "accept",
            "type": "string"
          },
          "expectedVersion": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "action",
          "expectedVersion"
        ],
        "type": "object"
      },
      "ActorResponseDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "system.manage",
                "tasks.view",
                "tasks.create",
                "tasks.accept",
                "tasks.complete",
                "tasks.review",
                "tasks.close",
                "demo.reset"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "role": {
            "example": "user",
            "type": "string"
          },
          "roleLabel": {
            "example": "演示用户",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "role",
          "roleLabel"
        ],
        "type": "object"
      },
      "AddTaskCommentDto": {
        "properties": {
          "content": {
            "example": "已完成第一阶段，明日提交结果",
            "maxLength": 1000,
            "minLength": 1,
            "type": "string"
          },
          "expectedVersion": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "content",
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "ApiErrorBodyDto": {
        "properties": {
          "code": {
            "example": "CONFLICT",
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "message": {
            "example": "任务已被其他操作更新",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorResponseDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          },
          "path": {
            "example": "/api/v1/tasks/task-1/actions",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "error",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
...
      "CreateTaskDto": {
        "properties": {
          "description": {
            "example": "绘制安全路线并带回现场印记",
            "maxLength": 500,
            "type": "string"
          },
          "dueDate": {
            "example": "2026-09-10",
            "format": "date",
            "type": "string"
          },
          "reward": {
            "example": "30 金币 · 稀有材料包",
            "maxLength": 120,
            "type": "string"
          },
          "title": {
            "example": "寻找遗失的地图碎片",
            "maxLength": 80,
            "type": "string"
          },
          "type": {
            "enum": [
              "exploration",
              "collection",
              "escort",
              "bounty",
              "building"
            ],
            "example": "exploration",
            "type": "string"
          }
        },
        "required": [
          "title",
          "type",
          "description",
          "reward",
          "dueDate"
        ],
        "type": "object"
      },
      "DeleteTaskCommentDto": {
        "properties": {
          "expectedVersion": {
            "example": 2,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "type": "object"
      },
      "EditTaskCommentDto": {
        "properties": {
          "content": {
            "example": "第一阶段已经完成，等待复核",
            "maxLength": 1000,
            "minLength": 1,
            "type": "string"
          },
          "expectedVersion": {
            "example": 2,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "content",
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "RenewExpiredTaskDto": {
        "properties": {
          "dueDate": {
            "example": "2026-09-10",
            "format": "date",
            "type": "string"
          },
          "expectedVersion": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "recoveryStrategy": {
            "enum": [
              "preserve_status",
              "reopened"
            ],
            "example": "preserve_status",
            "type": "string"
          }
        },
        "required": [
          "dueDate",
          "recoveryStrategy",
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "TaskResponseDto": {
        "properties": {
          "assignee": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ActorResponseDto"
              }
            ],
            "nullable": true,
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "dueDate": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "publisher": {
            "$ref": "#/components/schemas/ActorResponseDto"
          },
          "reward": {
            "type": "string"
          },
          "status": {
            "enum": [
              "not_started",
              "in_progress",
              "completed",
              "reopened",
              "closed",
              "expired"
            ],
            "type": "string"
          },
          "statusLabel": {
            "type": "string"
          },
          "timeline": {
            "items": {
              "discriminator": {
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TaskActivityResponseDto"
                },
                {
                  "$ref": "#/components/schemas/TaskCommentResponseDto"
                }
              ]
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "enum": [
              "exploration",
              "collection",
              "escort",
              "bounty",
              "building"
            ],
            "type": "string"
          },
          "typeLabel": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "workflowStatus": {
            "enum": [
              "not_started",
              "in_progress",
              "completed",
              "reopened",
              "closed"
            ],
            "type": "string"
          },
          "workflowStatusLabel": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "typeLabel",
          "description",
          "reward",
          "dueDate",
          "publisher",
          "assignee",
          "workflowStatus",
          "workflowStatusLabel",
          "status",
          "statusLabel",
          "createdAt",
          "updatedAt",
          "version",
          "timeline"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "demo-user": {
        "in": "header",
        "name": "X-Demo-User-Id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "演示身份与任务状态机的 demo-only REST 契约",
    "title": "告示牌 Noticeboard API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "listTasks",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "createTask",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}": {
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}/actions": {
      "post": {
        "operationId": "actOnTask",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}/comments": {
      "post": {
        "operationId": "createTaskComment",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTaskCommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}/comments/{commentId}": {
      "delete": {
        "operationId": "deleteTaskComment",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTaskCommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "editTaskComment",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTaskCommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{taskId}/expiration-renewal": {
      "post": {
        "operationId": "renewExpiredTask",
        "parameters": [
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewExpiredTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "tasks"
        ]
      }
    }
  },
  "servers": [],
...
}